Remove smart hotkeys and use only letters
By fixing the subclasses method I found that there were too many
statements to fit in a 0-9 hotkey list, so I replaced the numbers with
single letters. And because the letters are now used, I had to remove
the previous smart code that used the class name to figure out a good
hotkey.